moves Notification from teleterm to shared
