+++ packages/teleterm/src/ui/components/Notifcations/index.ts
@@ -1,3 +1,2 @@
 export * from './Notifications';
 export * from './NotificationsHost';
-export * from './types';
