--- packages/teleterm/src/ui/components/Notifcations/types.ts
@@ -1,9 +1,0 @@
-export interface NotificationItem {
-  content: NotificationItemContent;
-  severity: 'info' | 'warn' | 'error';
-  id: string;
-}
-
-export type NotificationItemContent =
-  | string
-  | { title: string; description: string };
