+++ packages/teleterm/src/ui/services/notifications/notificationsService.ts
@@ -1,10 +1,11 @@
 import { useStore } from 'shared/libs/stores';
 
-import { ImmutableStore } from 'teleterm/ui/services/immutableStore';
 import {
   NotificationItem,
   NotificationItemContent,
-} from 'teleterm/ui/components/Notifcations';
+} from '@gravitational/shared/components/Notification';
+
+import { ImmutableStore } from 'teleterm/ui/services/immutableStore';
 import { unique } from 'teleterm/ui/utils/uid';
 
 export class NotificationsService extends ImmutableStore<NotificationItem[]> {
